Refactor Code (#8)
* rename command files

* fix config command package and variable names

* update circleci client and utils structure

* [1/2] refactor code and project structure

* rename filename to make it more clear

* make golangci-lint happy

* delete unneccessary dir

* move auto complete functions to their respective command's files

* fix goimports errors

* move utility functions to existing packages

* fix spellings and case

* move commands and everything else to plugin package

* make variable names similar to as across whole package

* fix import order to follow best practices

* remove unneccessary files and add code for webhook functionality in single file

* Renamed some things for consistency

* Move build informations to webhook

* Separate httpinto multiple files

* Fix build

* I shouldn't code in a text editor

* Fix lint

* Fix

* I hate CRLF
