+++ server/.gitignore
@@ -1,2 +1,3 @@
 coverage.txt
 dist
+manifest.go
