--- server/circle/utils.go
@@ -1,7 +1,0 @@
-package circle
-
-import "fmt"
-
-func GetFullNameFromOwnerAndRepo(owner string, repository string) string {
-	return fmt.Sprintf("%s/%s", owner, repository)
-}
