+++ server/plugin/account.go
@@ -1,4 +1,4 @@
-package main
+package plugin
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 	"github.com/mattermost/mattermost-server/v5/model"
 
 	"github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle"
+	v1 "github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle/v1"
 )
 
 const (
@@ -23,6 +24,21 @@ const (
 	accountDisconnectTrigger  = "disconnect"
 	accountDisconnectHelpText = "Disconnect your Mattermost account from CircleCI"
 )
+
+func getAccountAutoCompleteData() *model.AutocompleteData {
+	account := model.NewAutocompleteData(accountTrigger, accountHint, accountHelpText)
+
+	view := model.NewAutocompleteData(accountViewTrigger, "", AccountViewHelpText)
+	connect := model.NewAutocompleteData(accountConnectTrigger, accountConnectHint, accountConnectHelpText)
+	connect.AddTextArgument("Generate a Personal API Token from your CircleCI user settings", accountConnectHint, "")
+	disconnect := model.NewAutocompleteData(accountDisconnectTrigger, "", accountDisconnectHelpText)
+
+	account.AddCommand(view)
+	account.AddCommand(connect)
+	account.AddCommand(disconnect)
+
+	return account
+}
 
 func (p *Plugin) executeAccount(args *model.CommandArgs, circleciToken string, split []string) (*model.CommandResponse, *model.AppError) {
 	subcommand := "help"
@@ -49,13 +65,13 @@ func (p *Plugin) executeAccount(args *model.CommandArgs, circleciToken string, s
 }
 
 func (p *Plugin) executeAccountView(args *model.CommandArgs, token string) (*model.CommandResponse, *model.AppError) {
-	user, ok := p.getCircleUserInfo(token)
-	if !ok {
+	user, err := v1.GetCircleUserInfo(token)
+	if err != nil {
 		p.API.LogInfo("Unable to get CircleCI info", "MM UserID", args.UserId)
 		return p.sendEphemeralResponse(args, errorConnectionText), nil
 	}
 
-	projects, _ := p.getCircleciUserProjects(token)
+	projects, _ := v1.GetCircleciUserProjects(token)
 	projectsListString := ""
 	for _, project := range projects {
 		// TODO : add circleCI url
@@ -68,8 +84,8 @@ func (p *Plugin) executeAccountView(args *model.CommandArgs, token string) (*mod
 		[]*model.SlackAttachment{
 			{
 				ThumbURL: user.AvatarURL,
-				Fallback: "User:" + circleciUserToString(user) + ". Email:" + *user.SelectedEmail,
-				Pretext:  "Information for CircleCI user " + circleciUserToString(user),
+				Fallback: "User:" + v1.CircleciUserToString(user) + ". Email:" + *user.SelectedEmail,
+				Pretext:  "Information for CircleCI user " + v1.CircleciUserToString(user),
 				Fields: []*model.SlackAttachmentField{
 					{
 						Title: "Name",
@@ -100,12 +116,12 @@ func (p *Plugin) executeAccountConnect(args *model.CommandArgs, split []string) 
 	}
 
 	if token, exists := p.Store.GetTokenForUser(args.UserId); exists {
-		user, ok := p.getCircleUserInfo(token)
-		if !ok {
+		user, err := v1.GetCircleUserInfo(token)
+		if err != nil {
 			return p.sendEphemeralResponse(args, "Internal error when reaching CircleCI"), nil
 		}
 
-		return p.sendEphemeralResponse(args, "You are already connected as "+circleciUserToString(user)), nil
+		return p.sendEphemeralResponse(args, "You are already connected as "+v1.CircleciUserToString(user)), nil
 	}
 
 	circleciToken := split[0]
