+++ server/plugin/project.go
@@ -1,4 +1,4 @@
-package main
+package plugin
 
 import (
 	"fmt"
@@ -6,6 +6,8 @@ import (
 
 	"github.com/jszwedko/go-circleci"
 	"github.com/mattermost/mattermost-server/v5/model"
+
+	v1 "github.com/nathanaelhoun/mattermost-plugin-circleci/server/circle/v1"
 )
 
 const (
@@ -22,6 +24,21 @@ const (
 	projectRecentBuildsHint     = "<username> <repository> <branch>"
 	projectRecentBuildsHelpText = "List the 10 last builds for a project"
 )
+
+func getProjectAutoComplete() *model.AutocompleteData {
+	project := model.NewAutocompleteData(projectTrigger, projectHint, projectHelpText)
+
+	projectList := model.NewAutocompleteData(projectListTrigger, projectListHint, projectListHelpText)
+	projectRecentBuild := model.NewAutocompleteData(projectRecentBuildsTrigger, projectRecentBuildsHint, projectRecentBuildsHelpText)
+	projectRecentBuild.AddTextArgument("Owner of the project's repository", "[username]", "")
+	projectRecentBuild.AddDynamicListArgument("", routeAutocompleteFollowedProjects, true)
+	projectRecentBuild.AddTextArgument("Branch name", "[branch]", "")
+
+	project.AddCommand(projectRecentBuild)
+	project.AddCommand(projectList)
+
+	return project
+}
 
 func (p *Plugin) executeProject(args *model.CommandArgs, circleciToken string, split []string) (*model.CommandResponse, *model.AppError) {
 	subcommand := "help"
@@ -45,8 +62,8 @@ func (p *Plugin) executeProject(args *model.CommandArgs, circleciToken string, s
 }
 
 func (p *Plugin) executeProjectList(args *model.CommandArgs, circleciToken string) (*model.CommandResponse, *model.AppError) {
-	projects, ok := p.getCircleciUserProjects(circleciToken)
-	if !ok {
+	projects, err := v1.GetCircleciUserProjects(circleciToken)
+	if err != nil {
 		return p.sendEphemeralResponse(args, errorConnectionText), nil
 	}
 
@@ -94,14 +111,14 @@ func (p *Plugin) executeProjectRecentBuilds(args *model.CommandArgs, circleciTok
 
 	text := "| Workflow | Job | Build | Subject | Start time | Status | Duration | Triggered by|\n| :---- | :----- | :----- | :----- | :----- | :----- | :---- | \n"
 	for _, build := range builds {
-		buildStartTime := buildStartTimeToString(build)
+		buildStartTime := v1.BuildStartTimeToString(build)
 
 		buildTime := "/"
 		if build.BuildTimeMillis != nil {
 			buildTime = strconv.Itoa(*build.BuildTimeMillis/1000) + "s"
 		}
 
-		statusImageMarkdown := buildStatusToMarkdown(build, p)
+		statusImageMarkdown := v1.BuildStatusToMarkdown(build, badgePassedURL, badgeFailedURL)
 
 		text += fmt.Sprintf("| % s | % s | [%d](%s) | `%s` | %s | %s | %s | %s |\n",
 			build.Workflows.WorkflowName,
